Post section completed for phtographer, videographer and singer

+++ ContentCreator/ContentCreator.Application/Common/DTOs/ResponseDTOs/FileTypeResponseModel.cs
@@ -5,8 +5,8 @@
         public Guid Id { get; set; }
         public string FileType { get; set; } = string.Empty;
         public string? FileExtension { get; set; }
-        public int? MinimumSize { get; set; }
-        public int? MaximumSize { get; set; }
+        public decimal? MinimumSize { get; set; }
+        public decimal? MaximumSize { get; set; }
         public bool IsActive { get; set; } = false;
     }
 }
